refactor : Reset winning rank before each test
