+++ src/test/java/domain/lotto/LottoGeneratorTest.java
