+++ src/test/java/domain/lotto/LottoTest.java
