+++ src/test/java/domain/lotto/PurchaseAmountTest.java
