+++ src/test/java/domain/statistics/RevenueCalculatorTest.java
@@ -1,15 +1,20 @@
-package lotto;
+package domain.statistics;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.within;
 
 import domain.lotto.PurchaseAmount;
-import domain.statistics.RevenueCalculator;
-import domain.statistics.WinningRank;
 import java.util.Arrays;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 public class RevenueCalculatorTest {
+
+    @BeforeEach
+    void set_up() {
+        Arrays.stream(WinningRank.values()).forEach(WinningRank::resetSuccessMatch);
+    }
 
     @AfterEach
     void 테스트_후_당첨_현황_초기화() {
@@ -20,7 +25,7 @@ public class RevenueCalculatorTest {
     void 최종_수익률_소수점_둘째자리까지_계산() {
         PurchaseAmount purchaseAmount = new PurchaseAmount("8000");
         WinningRank.match(3, false);
-
-        assertThat(RevenueCalculator.calculateRevenue(purchaseAmount)).isEqualTo(62.5);
+        
+        assertThat(RevenueCalculator.calculateRevenue(purchaseAmount)).isCloseTo(62.50, within(0.01));
     }
 }
