+++ src/test/java/domain/statistics/WinningStatisticsTest.java
@@ -1,14 +1,18 @@
-package lotto;
+package domain.statistics;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import domain.statistics.WinningRank;
-import domain.statistics.WinningStatistics;
 import java.util.Arrays;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 public class WinningStatisticsTest {
+
+    @BeforeEach
+    void set_up() {
+        Arrays.stream(WinningRank.values()).forEach(WinningRank::resetSuccessMatch);
+    }
 
     @AfterEach
     void 테스트_후_당첨_현황_초기화() {
