+++ src/test/java/domain/winningLotto/WinningLottoTest.java
